% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\docType{data}
\name{pdf_data}
\alias{pdf_data}
\alias{x_grid}
\title{Simulated PDF data.}
\format{
`pdf_data` is an n x p matrix, where n is the number of PDFs and p matches the length of x_grid.  x_grid contains the points at which the PDFs are evaluated (assumed equally spaced apart).
}
\usage{
pdf_data

x_grid
}
\description{
Data simulated using the the dfnWorks suite.
}
\details{
`pdf_data` is a data frame with 1,000 rows and 5 columns.
`x_grid`; is a timestamp of each of `full_data`'s 1,000 rows.
}
\examples{
pdf_data
x_grid
}
\keyword{datasets}
