% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_densities.R
\name{print.DeBoinR}
\alias{print.DeBoinR}
\title{Print function for a DeBoinR object.  Prints ggplot graphs and other output values.}
\arguments{
\item{x}{deboinr object. Fit from DeBoinR main method.}

\item{...}{Additional plotting arguments.}
}
\description{
Print function for a DeBoinR object.  Prints ggplot graphs and other output values.
}
